pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secaudit"
version = "0.2.0"
description = "Multi-repo GitHub security scanning tool with Claude-powered triage"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]
scanners = ["pip-audit>=2.7.0", "semgrep>=1.70.0"]

[project.scripts]
secaudit = "secaudit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
